Fixing wrong parameter types

+++ Software/Development/SemesterProject/App/Include/controller.h
@@ -13,6 +13,7 @@
 //
 #include "Constants.h"
 #include "PIController.h"
+#include "DSP28x_Project.h"
 
 float calculateIdReference(void);
 float calculateIqReference(float torqueReference);
@@ -23,7 +24,7 @@ float calculateIqReference(float torqueReference);
 void initPIControllers(void);
 float PiCalculationIQ(float reference, float measuredValue);
 float PiCalculationID(float reference, float measuredValue);
-float PiCalcualtionSpeed(float reference, float measuredValue);
+float PiCalcualtionSpeed(float reference, int16 measuredValue);
 void resetIntegrators(void);
 void resetIdIntegrator(void);
 void resetIqIntegrator(void);
